+++ tests/assets/mock_functional_experiment.py
@@ -1,20 +1,34 @@
 """
-This is the description
+This is the description of the module
 """
+import typing as t
 from pycomex.utils import folder_path, file_namespace
 from pycomex.functional.experiment import Experiment
 
-PARAMETER = 'experiment'
+# testing comment - do not remove
 
+# :param PARAMETER:
+#       This is a parameter description. If this description is given in the particular 
+#       format, it should technically show up in the parameter
+PARAMETER: t.Optional[str] = 'experiment'
 
+# :param PARAMETER2:
+#       This is a nother description
+PARAMETER2: str = 'hello world'
+
+print(__annotations__)
 @Experiment(base_path=folder_path(__file__),
             namespace=file_namespace(__file__),
             glob=globals(),
             debug=True)
-def experiment(e: Experiment):
+def experiment(e: Experiment) -> None:
+    # Some random comment that is not important
     e.log('starting experiment...')
-
+    # :hook hook:
+    #       This is a hook description. If given in this format, it should technically be discoverable 
+    #       and added to the typing dict.
     e.apply_hook('hook', parameter=10)
+    # This is some random other string that is not important.
     e['metrics/parameter'] = e.PARAMETER
     e.log(e['metrics/parameter'])